refactor: LogResult를 이용한 테스트 코드로 수정한다
